# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Emerj Pro API",
        "description": "API Documentation of Emerj Pro System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions",
                "operationId": "9b50e29e364179ebde24ac5fc59cd442",
                "responses": {
                    "200": {
                        "description": "List of transactions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "operationId": "54eaa3952808f639d46f969f4c8b98c5",
                "requestBody": {
                    "description": "Transaction data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "transaction_type_id",
                                    "payment_method_id",
                                    "transaction_ref",
                                    "status",
                                    "currency",
                                    "amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "transaction_type_id": {
                                        "description": "Transaction Type ID",
                                        "type": "integer"
                                    },
                                    "payment_method_id": {
                                        "description": "Payment Method ID",
                                        "type": "integer"
                                    },
                                    "transaction_ref": {
                                        "description": "Transaction reference",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Transaction status",
                                        "type": "boolean"
                                    },
                                    "currency": {
                                        "description": "Transaction currency",
                                        "type": "string"
                                    },
                                    "amount": {
                                        "description": "Transaction amount",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "user_id": [
                                                    "The user_id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt_auth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}